--- oshdb-tool/updater/src/main/java/org/heigit/bigspatialdata/updater/util/ReplicationFile.java
@@ -1,15 +1,0 @@
-package org.heigit.bigspatialdata.updater.util;
-
-import java.io.File;
-
-import org.openstreetmap.osmosis.replication.common.ReplicationState;
-
-public class ReplicationFile {
-	public final ReplicationState state;
-	public final File file;
-
-	public ReplicationFile(ReplicationState state, File file) {
-		this.state = state;
-		this.file = file;
-	}
-}
