+++ oshdb-tool/updater/src/main/java/org/heigit/bigspatialdata/updater/util/replication/XmlChangeReaderIterator.java
@@ -1,27 +1,48 @@
-package org.heigit.bigspatialdata.updater.util;
+package org.heigit.bigspatialdata.updater.util.replication;
 
+import org.heigit.bigspatialdata.updater.util.replication.DefaultChangeSink;
 import java.io.InputStream;
 import javax.xml.stream.FactoryConfigurationError;
 import javax.xml.stream.XMLInputFactory;
 import javax.xml.stream.XMLStreamException;
 import javax.xml.stream.XMLStreamReader;
+import org.heigit.bigspatialdata.updater.util.IteratorTmpl;
 import org.openstreetmap.osmosis.core.container.v0_6.ChangeContainer;
 import org.openstreetmap.osmosis.xml.v0_6.impl.OsmChangeHandler;
 import org.xml.sax.helpers.AttributesImpl;
 
-public class XmlChangeReaderIterator extends IteratorTmpl<ChangeContainer> implements DefaultChangeSink {
-  private final XMLStreamReader reader;
-  private final OsmChangeHandler osmHandler;
+/**
+ * Reads Changes from a XML-ReplicationFile.
+ */
+public class XmlChangeReaderIterator extends IteratorTmpl<ChangeContainer> implements
+    DefaultChangeSink {
+
   private ChangeContainer change = null;
+  private final OsmChangeHandler osmHandler;
+  private final XMLStreamReader reader;
 
   private XmlChangeReaderIterator(XMLStreamReader reader) {
     this.reader = reader;
     this.osmHandler = new OsmChangeHandler(this, true);
   }
 
+  /**
+   * Parse an XML-Replication-File into Changes.
+   *
+   * @param inputStream The data-stream
+   * @return An Iterator over ChangeContainers in that input
+   * @throws XMLStreamException
+   * @throws FactoryConfigurationError
+   */
   public static XmlChangeReaderIterator of(InputStream inputStream)
       throws XMLStreamException, FactoryConfigurationError {
-    return new XmlChangeReaderIterator(XMLInputFactory.newInstance().createXMLStreamReader(inputStream));
+    return new XmlChangeReaderIterator(XMLInputFactory.newInstance()
+        .createXMLStreamReader(inputStream));
+  }
+
+  @Override
+  public void process(ChangeContainer change) {
+    this.change = change;
   }
 
   @Override
@@ -42,9 +63,12 @@ public class XmlChangeReaderIterator extends IteratorTmpl<ChangeContainer> imple
           int attributeCount = reader.getAttributeCount();
           attributes.clear();
           for (int i = 0; i < attributeCount; i++) {
-            attributes.addAttribute(reader.getAttributeNamespace(i), reader.getAttributeLocalName(i),
-                reader.getAttributeName(i).toString(), reader.getAttributeType(i),
-                reader.getAttributeValue(i));
+            attributes
+                .addAttribute(reader.getAttributeNamespace(i),
+                    reader.getAttributeLocalName(i),
+                    reader.getAttributeName(i).toString(),
+                    reader.getAttributeType(i),
+                    reader.getAttributeValue(i));
           }
           osmHandler.startElement(uri, localName, qName, attributes);
           break;
@@ -65,8 +89,4 @@ public class XmlChangeReaderIterator extends IteratorTmpl<ChangeContainer> imple
     return change;
   }
 
-  @Override
-  public void process(ChangeContainer change) {
-    this.change = change;
-  }
 }
