+++ oshdb-util/src/main/java/org/heigit/bigspatialdata/oshdb/util/update/UpdateDbHelper.java
@@ -13,11 +13,23 @@ import org.heigit.bigspatialdata.oshdb.osm.OSMType;
 import org.heigit.bigspatialdata.oshdb.util.TableNames;
 import org.roaringbitmap.longlong.LongBitmapDataProvider;
 
+/**
+ * Helper class for the update-db.
+ */
 public class UpdateDbHelper {
 
   private UpdateDbHelper() {
   }
 
+  /**
+   * Get the bitmaps of updated entities from the update-db.
+   *
+   * @param dbBit connection to the bitmap-db. Commonly equals the connection to the update-db
+   * @return the current bitmaps where updated OSHEntites are flagged.
+   * @throws SQLException
+   * @throws IOException
+   * @throws ClassNotFoundException
+   */
   public static Map<OSMType, LongBitmapDataProvider> getBitMap(Connection dbBit) throws SQLException,
       IOException, ClassNotFoundException {
     Map<OSMType, LongBitmapDataProvider> map = new HashMap<>(3);
